Conflicts:
	models/core_updates.php
	views/ru/main.tpl

+++ models/core_updates.php
@@ -6,11 +6,13 @@
 	require_once 'core_queue.php';
 	require_once 'core_page.php';
 
+	require_once 'core_parser.php';
 	require_once 'core_comparator.php';
 
 	define('UPKIND_SIZE', 0);
 	define('UPKIND_GROUP', 1);
 	define('UPKIND_DELETE', 2);
+	define('UPKIND_INLINE', 3);
 
 	class AuthorWorker {
 		public function check($author_id) {
@@ -18,24 +20,15 @@
 			$a = $g->get($author_id);
 			preg_match('/^[\/]*(.*?)[\/]*$/', $a['link'], $link);
 			$link = $link[1] . '/';
-			$url  = 'http://samlib.ru/' . $link;
-//			$hash = base64_encode($link);
-//			$store = SUB_DOMEN . '/cache/' . $hash;
-
-//			if (!is_file($store)) {
-				$html = url_get_contents($url);
-//				if ($html !== false)
-//					file_put_contents($store, gzcompress/**/($html));
-//			} else
-//				$html = gzuncompress/**/(file_get_contents($store));
-
-			$html = mb_convert_encoding($html, 'UTF-8', 'CP1251');
-			if (($html === false) || trim($html) == '') {
+			$data = SAMLIBParser::getData($link, 'author');
+			if ($data === false) {
 				echo Loc::lget('samlib_down');
+				return false;
+			} elseif ($data === null) {
+				echo Loc::lget('parse_method_unknown');
 				return false;
 			}
 
-			$data = $this->parseAuthor($html);
 			$groups = $data[0];
 			$inline = $data[1];
 			$links = $data[2];
@@ -53,13 +46,16 @@
 
 			// check updates
 			$ga = GroupsAggregator::getInstance();
-			$d = $ga->fetch(array('filter' => '`author` = ' . $author_id, 'nocalc' => 1, 'desc' => 0, 'collumns' => '`id`, `group`, `title`'));
+			$d = $ga->fetch(array('filter' => '`author` = ' . $author_id, 'nocalc' => 1, 'desc' => 0, 'collumns' => '`id`, `group`, `title`, `link`'));
 			$gii = array();
+			$gl = array();
 			$gt = array();
 			if ($d['total'])
 				foreach ($d['data'] as $row) {
+					$g_id = intval($row['id']);
 					$gt = array_merge($gt, explode('@', $row['title']));
-					$gii[intval($row['id'])] = array(0 => $row['group'], 1 => $row['title']);
+					$gii[$g_id] = array(0 => $row['group'], 1 => $row['title']);
+					$gl[$g_id] = $row['link'];
 				}
 //			debug($gt);
 
@@ -75,10 +71,7 @@
 			}
 
 
-//				debug($g);
-
 			$new = array_diff($g, $gt);
-//			debug($new);
 			if (count($new)) {
 				echo Loc::lget('groups_updated') . '<br />';
 				foreach ($new as $idx => $title) {
@@ -91,9 +84,25 @@
 					, 'description' => addslashes(htmlspecialchars($groups[$idx][2], ENT_QUOTES))
 					));
 					$gi[$gid] = $idx;
+					$gl[$gid] = $inline[$idx];
 				}
 			// ok, groups updated
 			}
+
+			$g = array();
+			foreach ($gi as $g_id => $idx) {
+				$dbhas = isset($gl[$g_id]) && $gl[$g_id];
+				$pghas = isset($inline[$idx]) && $inline[$idx];
+				if ($dbhas ^ $pghas) // !1 && 2 || 1 && !2
+					$g[$g_id] = $inline[$idx];
+			}
+
+			$ua = UpdatesAggregator::getInstance();
+			foreach ($g as $g_id => $link) {
+				$ga->update(array('link' => $link), $g_id);
+				$ua->changed($g_id, UPKIND_INLINE, $link && (strpos($link, '/') === false));
+			}
+
 
 //			debug($gi, 'gi');
 
@@ -107,7 +116,6 @@
 			if (count($check)) { // ok, there are smth to check
 				$diff = array();
 				$pa = PagesAggregator::getInstance();
-				$ua = UpdatesAggregator::getInstance();
 				$d = $pa->fetch(array(
 					'filter' => '`author` = ' . $author_id . ' and (`link` in ("' . join('", "', array_keys($check)) . '"))'
 				, 'nocalc' => 1, 'desc' => 0
@@ -162,9 +170,90 @@
 				echo Loc::lget('no_updates') . '<br />';
 		}
 
+		function checkGroup($group_id) {
+			$ga = GroupsAggregator::getInstance();
+			$g = $ga->get($group_id, '`id`, `link`, `author`');
+			if ($g['id'] != $group_id) return false;
+			$author_id = intval($g['author']);
+
+			$aa = AuthorsAggregator::getInstance();
+			$a = $aa->get($author_id, '`link`');
+			preg_match('/^[\/]*(.*?)[\/]*$/', $a['link'], $link);
+			$link = $link[1] . '/' . $g['link'];
+			$data = SAMLIBParser::getData($link, 'group');
+			if ($data === false) {
+				echo Loc::lget('samlib_down');
+				return false;
+			} elseif ($data === null) {
+				echo Loc::lget('parse_method_unknown');
+				return false;
+			}
+
+			$links = &$data['links'];
+			$check = array();
+			// is there anything to check?
+			foreach ($links as $link => &$data) //            0      1        2
+				if ($data[1] > 0) // size > 0 ?   // link => {title, size, description}
+					$check[$link] = $data;
+
+			if (count($check)) { // ok, there are smth to check
+				$ua = UpdatesAggregator::getInstance();
+				$diff = array();
+				$pa = PagesAggregator::getInstance();
+				$d = $pa->fetch(array(
+					'filter' => '`author` = ' . $author_id . ' and (`link` in ("' . join('", "', array_keys($check)) . '"))'
+				, 'nocalc' => 1, 'desc' => 0
+				, 'collumns' => '`id`, `link`, `size`, `group`'
+				));
+				if ($d['total']) { // there are smth in db, check it if size diff
+					foreach ($d['data'] as $row) {
+						$data = $check[$link = $row['link']]; // recently parsed data about page at @link
+						unset($check[$link]); // don't check it later in this function
+
+						$page_id = $row['id'];
+						if (intval($row['size']) <> $data[1])
+							$diff[$page_id] = $link; // check it
+
+						$old_group = intval($row['group']);
+						if ($group_id != $old_group) {
+							echo patternize(Loc::lget('page_changed_group'), $row) . '<br />';
+							$pa->update(array('group' => $group_id), $page_id);
+							$ua->changed($page_id, UPKIND_GROUP, $old_group);
+						}
+					}
+				}
+
+				if (count($diff))
+					echo Loc::lget('pages_updated') . '<br />';
+
+				if (count($check)) {
+					echo Loc::lget('pages_new') . '<br />';
+					foreach ($check as $link => &$data) { // new pages
+						$pid = $pa->add(array(
+							'author' => $author_id
+						, 'group' => $group_id
+						, 'link' => $link
+						, 'title' => addslashes(htmlspecialchars($data[0], ENT_QUOTES))
+						, 'description' => addslashes(htmlspecialchars($data[2], ENT_QUOTES))
+						, 'size' => 0 // later it will be updated with diff checker
+						));
+						$diff[$pid] = $link;
+					}
+				}
+
+				if (count($diff)) {
+					echo Loc::lget('pages_queue') . '<br />';
+					$this->queuePages($author_id, $diff);
+				} else
+					echo Loc::lget('no_updates') . '<br />';
+			} else
+				echo Loc::lget('no_updates') . '<br />';
+		}
+
 		function queuePages($author, $pages) {
 			$a = AuthorsAggregator::getInstance();
-			$d = $a->get($author);
+			$d = $a->get($author, '`link`');
+			$alink = $d['link'];
 			$q = QueueAggregator::getInstance();
 
 			$worker_stamp = md5(uniqid(rand(),1));
@@ -175,7 +264,7 @@
 				echo Loc::lget('already_queued') . ":<br />";
 				foreach ($diff as $page) {
 					$link = $pages[$page];
-					echo "<a href=\"http://samlib.ru/$d[link]/$link\">$link</a><br />";
+					echo "<a href=\"http://samlib.ru/$alink/$link\">$link</a><br />";
 				}
 				echo '<br />';
 
@@ -184,7 +273,7 @@
 			echo Loc::lget('queued_pages') . ":<br />";
 			foreach ($idx as $page) {
 				$link = $pages[$page];
-				echo "<a href=\"http://samlib.ru/$d[link]/$link\">$link</a><br />";
+				echo "<a href=\"http://samlib.ru/$alink/$link\">$link</a><br />";
 			}
 		}
 
@@ -232,78 +321,49 @@
 					$pa->dbc->open();
 					/* - */
 
-					if (intval($row['size']) <> ($size = intval($size[2]))) {
-						$pa->update(array('size' => $size, 'time' => $time), $page);
-						$ua->changed($page, UPKIND_SIZE, $size - intval($row['size']));
-						echo " &nbsp;save to [/cache/pages/$page/$time.html]...<br />";
-						echo ' &nbsp;updated (' . $size . 'KB).<br />';
+					if ($size) {
+						if (intval($row['size']) <> ($size = intval($size[2]))) {
+							$pa->update(array('size' => $size, 'time' => $time), $page);
+							$ua->changed($page, UPKIND_SIZE, $size - intval($row['size']));
+							echo " &nbsp;save to [/cache/pages/$page/$time.html]...<br />";
+							echo ' &nbsp;updated (' . $size . 'KB).<br />';
+						}
+						$q->dbc->delete($q->TBL_DELETE, '`page` = ' . $page);
+					} else {
+						$q->dbc->update($q->TBL_INSERT, array('state' => 0, 'updated' => $time), '`id` = ' . $q_id);
+						echo Loc::lget('page_request_failed') . '<br />';
 					}
-					$q->dbc->delete($q->TBL_DELETE, '`page` = ' . $page);
 					$done++;
 					if (($timeout > 0) && (time() - $t > $timeout))
 						return $left - $done;
 				}
 				return 0;
-			}
+			} else
+				echo Loc::lget('nothing_to_update') . '<br />';
+
 			return 0;
 		}
 
-		function parseAuthor($html) {
-			$p1 = '- Блок ссылок на произведения -';
-			$p2 = '- Подножие -';
-			$i1 = strpos($html, $p1);
-			$i2 = strpos($html, $p2, $i1);
-			$m = substr($html, $i1, $i2 - $i1);
-			$m = preg_replace('/(<!-+|--+>)/', '', $m);
+		function groupsToUpdate($force = 0) {
+			$dbc = msqlDB::o();
+			$t = time() - ($force ? 5 : 60 * 30); // 30 minutes
+			$s = $dbc->select('groups'
+			, '`time` < ' . $t . ' and `link` <> "" and `link` not like "/%" order by `time`'
+			, '`id` as `0`'
+			);
+			$a = array();
+			if ($s)
+				foreach($dbc->fetchrows($s) as $row)
+					$a[] = intval($row[0]);
 
-			preg_match_all('/<a name="?gr(\d+)"?>(.*?)(<gr|<\/a)/i', $m, $g, PREG_SET_ORDER);
-			$groups = array();
-			$inline = array();
-			foreach($g as $match) {
-				$id = intval($match[1]);
-
-				$t = preg_replace('/((<[^>]+>)|([<\/]))/', '', $match[2]);
-				$t = ucfirst(preg_replace('/:$/', '', $t));
-				$groups[] = array($id, $t);
-
-				if (preg_match('/a href="?([^ ">]+)"?/i', $match[2], $t))
-					$inline[count($groups) - 1] = $t[1];
-			}
-
-//			debug($groups);
-//			debug($inline);
-
-			$pieces = explode('<a name=gr', $m);
-			unset($pieces[0]);
-
-			$links = array();
-			foreach ($groups as $idx => &$g) {
-				$id = $g[0];
-				$block = array_shift($pieces);
-				preg_match('/<br>(.*?)<(dl|\/small)>/ism', $block, $t1);
-				$g[2] = trim(preg_replace(array('/<br( \/)?>/i', '/<[^>]+>/'), array(PHP_EOL, ''), $t1[1]));
-//				debug($g[2]);
-				preg_match_all('/<dl>(.*?)<\/dl>/i', $block, $t);
-//				debug($t, htmlspecialchars($block));
-				foreach ($t[1] as $link) {
-					preg_match('/<a href="?([^ >"]+)"?>(.*?)<\/a>[^<]*<b>(.*?)k<\/b>/i', $link, $t2);
-					preg_match('/<dd>(.*)/i', $link, $t3);
-//					debug($t2, htmlspecialchars($link));
-					$links[$id][$t2[1]] = array($idx, strip_tags($t2[2]), intval($t2[3]), strip_tags($t3[1], '<br><p>'));
-				}
-//				debug($links[$id], "Group #$id");
-			}
-//			debug($m);
-//			debug($html);
-			preg_match('/<h3>(.*?)<br/i', $html, $t);
-			$t = preg_replace('/:$/', '', $t[1]);
-			return array($groups, $inline, $links, $t);
+			return $a;
 		}
+
 	}
 
 	class UpdatesAggregator extends Aggregator {
 		static $instance = null;
-		var $TBL_FETCH  = '`updates` u, `pages` p, `authors` a, `groups` g';
+		var $TBL_FETCH  = 'updates';
 		var $TBL_INSERT = 'updates';
 		var $TBL_DELETE = 'updates';
 		var $collumns = array(
@@ -324,20 +384,100 @@
 			return self::$instance;
 		}
 
-		function getUpdates($page) {
+		function getUpdates($limit) {
 			$d = $this->fetch(array(
-				'filter' => 'u.`page` = p.`id` and a.`id` = p.`author` and g.`id` = p.`group`'
-			, 'collumns' => 'p.`id`, p.`title`, u.`kind`, u.`value`, a.`fio`, p.`author`, g.`title` as `group_title`, p.`group`, u.`time`'
-			, 'desc' => true
-			, 'page' => $page
+				'desc' => true
+			, 'page' => $limit - 1
 			, 'pagesize' => $this->FETCH_PAGE
 			));
-			return $d;
+			$r = array();
+			if ($d['total']) {
+				$a = array();
+				$g = array();
+				$p = array();
+				foreach ($d['data'] as &$row) {
+					switch ($row['kind']) {
+					case UPKIND_INLINE: // group updates
+						$g[intval($row['page'])] = 1;
+						break;
+					case UPKIND_GROUP:
+						$g[intval($row['value'])] = 1;
+					default:
+						$p[intval($row['page'])] = 1;
+						break;
+					}
+				}
+				$aa = AuthorsAggregator::getInstance();
+				$pa = PagesAggregator::getInstance();
+				$ga = GroupsAggregator::getInstance();
+				$pd = $pa->get(array_keys($p), '`id`, `title`, `author`, `group`');
+				foreach ($pd as &$row) {
+					$p[intval($row['id'])] = $row;
+					$g[intval($row['group'])] = 1;
+					$a[intval($row['author'])] = 1;
+				}
+				$gd = $ga->get(array_keys($g), '`id`, `title`, `author`');
+				$g = array();
+				foreach ($gd as &$row) {
+					$g[intval($row['id'])] = $row;
+					$a[intval($row['author'])] = 1;
+				}
+				$ad = $aa->get(array_keys($a), '`id`, `fio`');
+				$a = array();
+				foreach ($ad as &$row)
+					$a[intval($row['id'])] = $row['fio'];
+
+				foreach ($d['data'] as &$row) {
+					switch ($kind = $row['kind']) {
+					case UPKIND_INLINE: // group updates
+						$gid = intval($row['page']);
+						$aid = $g[$gid]['author'];
+						$r[] = array('id' => $gid, 'kind' => $kind, 'value' => $row['value']
+						, 'group' => $gid
+						, 'group_title' => $g[$gid]['title']
+						, 'author' => $aid
+						, 'fio' => $a[$aid]
+						);
+						break;
+					case UPKIND_GROUP:
+						$pid = intval($row['page']);
+						$gid = $p[$pid]['group'];
+						$aid = $p[$pid]['author'];
+						$goid = intval($row['value']);
+						$r[] = array('id' => $gid, 'kind' => $kind, 'value' => $row['value']
+						, 'page' => $pid
+						, 'title' => $p[$pid]['title']
+						, 'group' => $gid
+						, 'group_title' => $g[$gid]['title']
+						, 'old_id' => $goid
+						, 'old_title' => ($t = uri_frag($g, $goid, 0, 0)) ? $t['title'] : '&lt;!&gt;'
+						, 'author' => $aid
+						, 'fio' => $a[$aid]
+						);
+						break;
+					default: // page updates
+						$pid = intval($row['page']);
+						$aid = $p[$pid]['author'];
+						$gid = $p[$pid]['group'];
+						$r[] = array('id' => $pid, 'kind' => $kind, 'value' => $row['value']
+						, 'page' => $pid
+						, 'title' => $p[$pid]['title']
+						, 'group' => $gid
+						, 'group_title' => $g[$gid]['title']
+						, 'author' => $aid
+						, 'fio' => $a[$aid]
+						);
+						break;
+					}
+				}
+			}
+			return array('data' => $r, 'total' => $d['total']);
 		}
 
 		function changed($page, $kind, $value) {
 			return $this->add(array('page' => $page, 'kind' => $kind, 'value' => $value));
 		}
+
 	}
 
 ?>
